requirements.txt
alembic==1.7.5
attrs==21.2.0
autopage==0.4.0
cliff==3.9.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.3.1
colorlog==6.6.0
cycler==0.11.0
fonttools==4.28.1
greenlet==1.1.2
gym==0.19.0
imageio==2.11.1
importlib-metadata==4.8.2
importlib-resources==5.4.0
jsonpickle==2.0.0
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
networkx==2.6.3
numpy==1.21.4
opencv-python==4.5.4.58
optuna==2.10.0
packaging==21.3
palettable==3.3.0
pandas==1.3.4
pbr==5.8.0
Pillow==8.4.0
prettytable==2.4.0
pygame==1.9.6
pyparsing==3.0.6
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
sb3-contrib==1.3.0
scikit-image==0.18.3
scipy==1.7.2
setuptools-scm==6.3.2
six==1.16.0
SQLAlchemy==1.4.27
stable-baselines3==1.3.0
stevedore==3.5.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
tqdm==4.62.3
typing_extensions==4.0.0
wcwidth==0.2.5
zipp==3.6.0
